Make the 'memConstEqual' function run in constant time.
There was a possible timing side channel in the old code, since it used
a (&&!) function that dispatched on boolean parameters the execution time
of that function might differ for different values of that parameter
(the compiler will probably implement (&&!) with branching).

This will of course depend on how the compiler chooses to implement the
(&&!) function.

This version instead does only constant-time operations on the two input
strings. The code is similar to the classic C version of the constant time
comparison algorithm.

```
int
constanttime_eq(char *p1, char *p2, size_t len)
{
    int i;
    char ret;
    for (i=0; i<len; i++) {
        ret |= p1[i] ^ p2[i];
    }
    return 0 == ret;
}
```